Moving to the non-preview version of the Container Instance API
